+++ src/cli/concat.rs
@@ -33,8 +33,10 @@ impl<'a> ConcatParser<'a> {
     pub(in crate::cli) fn concat(&mut self) {
         let input_fmt = self.parse_input_fmt(self.matches);
         let datatype = self.parse_datatype(self.matches);
-        let output = self.parse_output(self.matches);
         let output_fmt = self.parse_output_fmt(self.matches);
+        let prefix = self.parse_output(self.matches);
+        let fname = self.create_fname(&prefix, &output_fmt);
+        let output = prefix.join(fname);
         let part_fmt = self.parse_partition_fmt(self.matches);
         self.check_partition_format(&output_fmt, &part_fmt);
         let task_desc = "Alignment concatenation";
@@ -52,6 +54,8 @@ impl<'a> ConcatParser<'a> {
             &input_fmt,
             &datatype,
         );
+
+        check_output_dir_exist(&prefix);
         let concat = msa::MSAlignment::new(&input_fmt, &output, &output_fmt, &part_fmt);
 
         concat.concat_alignment(&mut files, &datatype);