+++ src/cli/id.rs
@@ -16,23 +16,14 @@ impl InputCli for IdParser<'_> {
     }
 }
 
-impl OutputCli for IdParser<'_> {
-    fn parse_output(&self, matches: &ArgMatches) -> PathBuf {
-        if matches.is_present("output") {
-            let output = self.parse_output(matches);
-            output.with_extension("txt")
-        } else {
-            let input = self.parse_dir_input(matches);
-            input.with_extension("txt")
-        }
-    }
-}
+impl OutputCli for IdParser<'_> {}
 
 impl InputPrint for IdParser<'_> {}
 
 pub(in crate::cli) struct IdParser<'a> {
     matches: &'a ArgMatches<'a>,
     input_dir: Option<PathBuf>,
+    output: PathBuf,
 }
 
 impl<'a> IdParser<'a> {
@@ -40,6 +31,7 @@ impl<'a> IdParser<'a> {
         Self {
             matches,
             input_dir: None,
+            output: PathBuf::new(),
         }
     }
 
@@ -54,6 +46,7 @@ impl<'a> IdParser<'a> {
         } else {
             self.parse_input_wcard(self.matches)
         };
+
         self.print_input_multi(
             &self.input_dir,
             task_desc,
@@ -61,31 +54,49 @@ impl<'a> IdParser<'a> {
             &input_fmt,
             &datatype,
         );
+        self.output = self.parse_output(self.matches).with_extension("txt");
+        check_output_file_exist(&self.output);
         let spin = utils::set_spinner();
         spin.set_message("Indexing IDs..");
         let ids = IDs::new(&files, &input_fmt, &datatype).get_id_all();
         spin.finish_with_message("DONE!");
-        self.write_results(&ids);
+        self.write_results(&ids).expect("Failed writing results");
+        self.print_output(ids.len());
     }
 
     fn is_input_dir(&self) -> bool {
         self.matches.is_present("dir")
     }
 
-    fn write_results(&self, ids: &IndexSet<String>) {
-        let fname = self.parse_output(self.matches);
-        let file = File::create(&fname).expect("CANNOT CREATE AN OUTPUT FILE");
+    fn write_results(&self, ids: &IndexSet<String>) -> Result<()> {
+        let file = File::create(&self.output).expect("CANNOT CREATE AN OUTPUT FILE");
         let mut writer = BufWriter::new(file);
         ids.iter().for_each(|id| {
             writeln!(writer, "{}", id).unwrap();
         });
-        writer.flush().unwrap();
-        self.print_output(&fname, ids.len());
+        writer.flush()?;
+        Ok(())
     }
 
-    fn print_output(&self, output: &Path, ids: usize) {
+    fn print_output(&self, ids: usize) {
         log::info!("\n{}", Yellow.paint("Output"));
         log::info!("{:18}: {}", "Total unique IDs", ids);
-        log::info!("{:18}: {}", "File output", output.display());
+        log::info!("{:18}: {}", "File output", self.output.display());
     }
 }
+
+// #[cfg(test)]
+// mod test {
+//     use super::*;
+//     use clap::{App, Arg};
+
+//     #[test]
+//     fn get_id_output_path_test() {
+//         let arg = App::new("segul-test")
+//             .arg(Arg::with_name("dir").default_value("./test_dir/"))
+//             .get_matches();
+//         let id = IdParser::new(&arg);
+//         let res = PathBuf::from("./test_dir.txt");
+//         assert_eq!(res, id.output);
+//     }
+// }